Revamp provider/model selection
