+++ src/renderer/src/lib/types.ts
@@ -1,5 +1,6 @@
 import type {UIMessage} from 'ai';
 import {z} from "zod";
+import {ModelProviderType} from "@/common/models/modelProvider";
 
 export const messageMetadataSchema = z.object({
     createdAt: z.string(),
@@ -27,3 +28,22 @@ export interface Attachment {
     url: string;
     contentType: string;
 }
+
+export const ProviderInfo: Record<ModelProviderType, { name: string; description: string }> = {
+    [ModelProviderType.OPENAI]: {
+        name: 'OpenAI',
+        description: 'Access state-of-the-art models like GPT-4 and GPT-3.5.'
+    },
+    [ModelProviderType.ANTHROPIC]: {
+        name: 'Anthropic',
+        description: 'Utilize the Claude family of models, known for safety and performance.'
+    },
+    [ModelProviderType.GOOGLE]: {
+        name: 'Google',
+        description: 'Leverage the powerful Gemini family of models from Google AI.'
+    },
+    [ModelProviderType.CUSTOM]: {
+        name: 'Custom',
+        description: 'Connect to any OpenAI-compatible API endpoint.'
+    }
+};
